BUG FIX in bb = MaskApi.toBbox(R) !!!
Details: this bug only affected extraction of bounding boxes for RLEs that took full image height. In this case the returned bb should have y=0 and box_height=image_height. However, the bug would have the y values incorrect for such boxes (the extraction of the bbox would miss the fact that the RLE encoded mask loops from y=h-1 to y=0 in certain cases). 

An example (in Matlab) of the bug before the fix is as follows:
 M=uint8([0 1; 1 1]); bbox=MaskApi.toBbox(MaskApi.encode(M))
This would return [0 1 2 1] instead of [0 0 2 2]. This has now been corrected.